+++ BoletoBusMonolitic.Web/Data/Context/BoletoBusContext.cs
@@ -1,4 +1,4 @@
-﻿namespace BoletoBusMonolitic.Web.Date.Context
+﻿namespace BoletoBusMonolitic.Web.Data.Context
 {
     public class BoletoBusContext
     {
