+++ BoletoBusMonolitic.Web/Data/Core/BaseEntity.cs
@@ -1,0 +1,6 @@
+﻿namespace BoletoBusMonolitic.Web.Data.Core
+{
+    public class BaseEntity
+    {
+    }
+}
