+++ BoletoBusMonolitic.Web/Data/Entites/Viaje.Cs
@@ -1,0 +1,9 @@
+﻿using BoletoBusMonolitic.Web.Data.Core;
+
+namespace BoletoBusMonolitic.Web.Data.Entites;
+
+
+public class Viaje : BaseEntity
+{
+
+}
