+++ BoletoBusMonolitic.Web/Data/Interfaces/IViaje.cs
@@ -1,0 +1,9 @@
+﻿using BoletoBusMonolitic.Web.Data.Entites;
+
+namespace BoletoBusMonolitic.Web.Data.Interfaces
+{
+    public interface IViajeDb
+    {
+        List<Viaje> GetViajeList();
+    }
+}
