+++ BoletoBusMonolitic.Web/Data/Models/ViajeModel.Cs
@@ -1,0 +1,6 @@
+﻿namespace BoletoBusMonolitic.Web.Data.Models
+{
+    public class ViajeModel
+    {
+    }
+}
