--- BoletoBusMonolitic.Web/Date/Core/BaseEntity.cs
@@ -1,6 +1,0 @@
-﻿namespace BoletoBusMonolitic.Web.Date.Core
-{
-    public class BaseEntity
-    {
-    }
-}
