--- BoletoBusMonolitic.Web/Date/Entites/Viaje.Cs
@@ -1,9 +1,0 @@
-﻿using BoletoBusMonolitic.Web.Date.Core;
-
-namespace BoletoBusMonolitic.Web.Date.Entites;
-
-
-public class Viaje : BaseEntity
-{
-
-}
