--- BoletoBusMonolitic.Web/Date/Interfaces/IViaje.cs
@@ -1,9 +1,0 @@
-﻿using BoletoBusMonolitic.Web.Date.Entites;
-
-namespace BoletoBusMonolitic.Web.Date.Interfaces
-{
-    public interface IViajeDb
-    {
-        List<Viaje> GetViajeList();
-    }
-}
