--- BoletoBusMonolitic.Web/Date/Models/ViajeModel.Cs
@@ -1,6 +1,0 @@
-﻿namespace BoletoBusMonolitic.Web.Date.Models
-{
-    public class ViajeModel
-    {
-    }
-}
